refactor: move UTXO listing to its own component
